+++ Sources/Entities/AttachmentArchiveItemReadable.swift
@@ -9,7 +9,7 @@ public struct AttachmentArchiveItemReadable: Codable {
     /// The path of the archive item.
     public var path: String?
     /// The position of the item within the archive.
-    public var index: Int?
+    public var index: Int64?
     /// The size of the archive item.
     public var size: String?
     /// The MIME type of the archive item.
@@ -17,7 +17,7 @@ public struct AttachmentArchiveItemReadable: Codable {
     /// The label for the archive item.
     public var label: String?
 
-    public init(path: String? = nil, index: Int? = nil, size: String? = nil, mediaType: String? = nil, label: String? = nil) {
+    public init(path: String? = nil, index: Int64? = nil, size: String? = nil, mediaType: String? = nil, label: String? = nil) {
         self.path = path
         self.index = index
         self.size = size
@@ -28,7 +28,7 @@ public struct AttachmentArchiveItemReadable: Codable {
     public init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: StringCodingKey.self)
         self.path = try values.decodeIfPresent(String.self, forKey: "path")
-        self.index = try values.decodeIfPresent(Int.self, forKey: "index")
+        self.index = try values.decodeIfPresent(Int64.self, forKey: "index")
         self.size = try values.decodeIfPresent(String.self, forKey: "size")
         self.mediaType = try values.decodeIfPresent(String.self, forKey: "mediaType")
         self.label = try values.decodeIfPresent(String.self, forKey: "label")
