+++ Sources/Entities/BulkProjectPermissionGrants.swift
@@ -9,11 +9,11 @@ public struct BulkProjectPermissionGrants: Codable {
     /// A project permission,
     public var permission: String
     /// IDs of the issues the user has the permission for.
-    public var issues: [Int]
+    public var issues: [Int64]
     /// IDs of the projects the user has the permission for.
-    public var projects: [Int]
+    public var projects: [Int64]
 
-    public init(permission: String, issues: [Int], projects: [Int]) {
+    public init(permission: String, issues: [Int64], projects: [Int64]) {
         self.permission = permission
         self.issues = issues
         self.projects = projects
@@ -22,8 +22,8 @@ public struct BulkProjectPermissionGrants: Codable {
     public init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: StringCodingKey.self)
         self.permission = try values.decode(String.self, forKey: "permission")
-        self.issues = try values.decode([Int].self, forKey: "issues")
-        self.projects = try values.decode([Int].self, forKey: "projects")
+        self.issues = try values.decode([Int64].self, forKey: "issues")
+        self.projects = try values.decode([Int64].self, forKey: "projects")
     }
 
     public func encode(to encoder: Encoder) throws {
