+++ Sources/Entities/BulkProjectPermissions.swift
@@ -7,13 +7,13 @@ import NaiveDate
 /// Details of project permissions and associated issues and projects to look up.
 public struct BulkProjectPermissions: Codable {
     /// List of issue IDs.
-    public var issues: [Int]?
+    public var issues: [Int64]?
     /// List of project IDs.
-    public var projects: [Int]?
+    public var projects: [Int64]?
     /// List of project permissions.
     public var permissions: [String]
 
-    public init(issues: [Int]? = nil, projects: [Int]? = nil, permissions: [String]) {
+    public init(issues: [Int64]? = nil, projects: [Int64]? = nil, permissions: [String]) {
         self.issues = issues
         self.projects = projects
         self.permissions = permissions
@@ -21,8 +21,8 @@ public struct BulkProjectPermissions: Codable {
 
     public init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: StringCodingKey.self)
-        self.issues = try values.decodeIfPresent([Int].self, forKey: "issues")
-        self.projects = try values.decodeIfPresent([Int].self, forKey: "projects")
+        self.issues = try values.decodeIfPresent([Int64].self, forKey: "issues")
+        self.projects = try values.decodeIfPresent([Int64].self, forKey: "projects")
         self.permissions = try values.decode([String].self, forKey: "permissions")
     }
